fix: bug in the PowerSpySensor
The division for getting the ratio between the elapsed and global times was wrong, it returned an Int.
It's now replaced by a double.

+++ src/main/scala/org/powerapi/module/powerspy/PowerSpySensor.scala
@@ -118,7 +118,7 @@ class PowerSpySensor(eventBus: MessageBus, osHelper: OSHelper, timeout: Timeout)
       case _ => 0l
     }
 
-    TargetUsageRatio(targetTime / globalTime)
+    TargetUsageRatio(targetTime.doubleValue / globalTime)
   }
 
   /**
